Cleaning up Main
