+++ src/app/Mastermind.java
@@ -1,0 +1,4 @@
+package app;
+
+public class Mastermind {
+}
